+++ src/example_app/example_app/documents/admin.py
@@ -45,6 +45,7 @@ class DocsAdmin(ResourceAdmin):
         'weight',
         'tags',
         'meta',
+        'nullable_comment',
     )
     list_view = ListViewSettings(
         fields=Exclude('data', 'image', 'meta.custom'),